backend/requirements.txt
Flask
Flask-CORS
Flask-JWT-Extended
Werkzeug
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
tensorflow>=2.13.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-dotenv
gunicorn
bcrypt
marshmallow
email-validator
phonenumbers
redis
celery
razorpay
google-generativeai>=0.3.0
requests>=2.31.0